# fix: eliminate Python validation race condition on Windows