--- a/test/python/wrapper.test.ts
+++ b/test/python/wrapper.test.ts
@@ -24,9 +24,8 @@ jest.mock('../../src/python/pythonUtils', () => {
     runPython: jest.fn(originalModule.runPython),
     tryPath: jest.fn(),
     getSysExecutable: jest.fn(),
-    state: {
-      cachedPythonPath: '/usr/bin/python3',
-    },
+    // Use the real state object so all implementations share the same cache
+    state: originalModule.state,
   };
 });
 interface TestResult {
@@ -88,32 +87,36 @@ describe('wrapper', () => {
     });
   });
   describe('validatePythonPath race conditions', () => {
-    let actualPythonUtils: any;
-
     beforeAll(() => {
-      // Get the actual module - we'll use the real validatePythonPath implementation
-      // but mock the low-level Python detection functions to avoid actual system calls
-      // that can hang on Windows.
-      actualPythonUtils = jest.requireActual('../../src/python/pythonUtils');
+      // Restore the real validatePythonPath implementation while keeping
+      // tryPath and getSysExecutable mocked to avoid actual system calls
+      const realModule = jest.requireActual('../../src/python/pythonUtils');
+      jest.mocked(validatePythonPath).mockImplementation(realModule.validatePythonPath);
     });
 
     beforeEach(() => {
       // Reset the cached path and validation promise before each test
-      actualPythonUtils.state.cachedPythonPath = null;
-      actualPythonUtils.state.validationPromise = null;
+      state.cachedPythonPath = null;
+      state.validationPromise = null;
 
-      // Configure the module-level mocks to return a valid Python path
-      jest.mocked(tryPath).mockResolvedValue('/usr/bin/python3');
-      jest.mocked(getSysExecutable).mockResolvedValue('/usr/bin/python3');
+      // Configure the module-level mocks with realistic delays to simulate Windows behavior
+      // This is critical for testing race conditions that only appear under slow I/O
+      jest.mocked(tryPath).mockImplementation(async (path: string) => {
+        // Simulate slow Windows process spawning and antivirus scanning
+        await new Promise((resolve) => setTimeout(resolve, 50));
+        return '/usr/bin/python3';
+      });
+      jest.mocked(getSysExecutable).mockImplementation(async () => {
+        // Simulate slow Windows 'where' command and py launcher
+        await new Promise((resolve) => setTimeout(resolve, 100));
+        return '/usr/bin/python3';
+      });
     });
     it('should handle concurrent validatePythonPath calls without race conditions', async () => {
-      const { validatePythonPath: realValidatePythonPath, state: realState } = actualPythonUtils;
-      // Force cache miss
-      realState.cachedPythonPath = null;
       // Launch multiple concurrent validations
       const concurrentCalls = 10;
       const promises = Array.from({ length: concurrentCalls }, (_, i) =>
-        realValidatePythonPath('python', false).then(
+        validatePythonPath('python', false).then(
           (result: string): TestResult => ({ testId: i, result }),
         ),
       );
@@ -125,19 +128,17 @@ describe('wrapper', () => {
         expect(typeof result.result).toBe('string');
       });
       // All results should be consistent (no race condition)
+      // If there was a race condition, different calls might get different results
       const uniqueResults = new Set(results.map((r: TestResult) => r.result));
       expect(uniqueResults.size).toBe(1);
       // Cache should be populated
-      expect(realState.cachedPythonPath).toBeTruthy();
+      expect(state.cachedPythonPath).toBeTruthy();
     }, 10000); // Increase timeout for this test
     it('should handle mixed explicit/implicit validation calls consistently', async () => {
-      const { validatePythonPath: realValidatePythonPath, state: realState } = actualPythonUtils;
-      // Force cache miss
-      realState.cachedPythonPath = null;
       // Create mixed explicit/implicit calls
       const mixedPromises = Array.from({ length: 8 }, (_, i) => {
         const isExplicit = i % 2 === 0;
-        return realValidatePythonPath('python', isExplicit).then(
+        return validatePythonPath('python', isExplicit).then(
           (result: string): MixedTestResult => ({
             testId: i,
             result,
@@ -162,19 +163,17 @@ describe('wrapper', () => {
       const uniqueExplicitResults = new Set(explicitResults);
       const uniqueImplicitResults = new Set(implicitResults);
       // Both explicit and implicit calls should return consistent results
+      // If there was a race condition, explicit and implicit calls might return different results
       expect(uniqueExplicitResults.size).toBe(1);
       expect(uniqueImplicitResults.size).toBe(1);
       // Explicit and implicit results should be the same
       expect(explicitResults[0]).toBe(implicitResults[0]);
     }, 10000);
     it('should handle rapid successive calls without race conditions', async () => {
-      const { validatePythonPath: realValidatePythonPath, state: realState } = actualPythonUtils;
-      // Force cache miss
-      realState.cachedPythonPath = null;
       // Launch rapid successive calls
       const rapidCalls = 20;
       const promises = Array.from({ length: rapidCalls }, (_, i) =>
-        realValidatePythonPath('python', false).then(
+        validatePythonPath('python', false).then(
           (result: string): TestResult => ({ testId: i, result }),
         ),
       );
@@ -186,10 +185,11 @@ describe('wrapper', () => {
         expect(typeof result.result).toBe('string');
       });
       // All results should be consistent
+      // If there was a race condition, different calls could get different cached values
       const uniqueResults = new Set(results.map((r: TestResult) => r.result));
       expect(uniqueResults.size).toBe(1);
       // Cache should be populated with the consistent result
-      expect(realState.cachedPythonPath).toBe(results[0].result);
+      expect(state.cachedPythonPath).toBe(results[0].result);
     }, 10000);
   });
 });